fix build error

+++ onert-micro/onert-micro/src/execute/kernels/SVDF.cpp
@@ -158,10 +158,10 @@ OMStatus onert_micro::execute::execute_kernel_CircleSVDF(const OMExecuteArgs &ex
     case circle::TensorType_INT8:
     {
       status = pal::SVDF(
-        utils::castInputData<int8_t>(input_data), utils::castInputData<int8_t>(weights_feature_data),
-        utils::castInputData<int8_t>(weights_time_data), utils::castInputData<int8_t>(bias_data),
-        utils::castOutputData<int8_t>(activation_state_data),
-        utils::castOutputData<int8_t>(scratch_buffer), utils::castOutputData<int8_t>(output_data),
+        utils::castInputData<float>(input_data), utils::castInputData<float>(weights_feature_data),
+        utils::castInputData<float>(weights_time_data), utils::castInputData<float>(bias_data),
+        utils::castOutputData<float>(activation_state_data),
+        utils::castOutputData<float>(scratch_buffer), utils::castOutputData<float>(output_data),
         rank, input_size, batch_size, num_filters, num_units, memory_size,
         options->fused_activation_function());
     }
